refactor(factory): delete sandbox fleet and base-checkpoint subsystem
The fleet's four sources of sandbox truth (persisted ids/workdirs, pools
and revival ladder, Platform DB status, provider state) are replaced by
provider-owned id-keyed getOrCreate: sandbox identity is the session id,
workdirs are computed deterministically, and setup runs inside the start
lifecycle. Deliberately dropped with the fleet: the per-replica sandbox
budget (SandboxBudgetError), transient git-failure retry classification,
and the cross-session sandbox pool. The still-live materialization types
(MaterializationSandbox, PrepareProgress) move to sandbox/materialization.
