+++ mastracode/factory/src/integrations/github/routes.ts
@@ -20,8 +20,7 @@ import type { FactoryStorage } from '@mastra/core/storage';
 import type { Context } from 'hono';
 import { streamSSE } from 'hono/streaming';
 import type { RouteAuth } from '../../routes/route.js';
-import { baseCheckpointIsStale } from '../../sandbox/base-checkpoint-triggers.js';
-import type { MaterializationSandbox, PrepareProgress, ProgressFn } from '../../sandbox/fleet.js';
+import type { MaterializationSandbox, PrepareProgress, ProgressFn } from '../../sandbox/materialization.js';
 import type { FactorySandboxRuntime } from '../../sandbox/session-sandbox.js';
 import { peekSessionSandbox } from '../../sandbox/session-sandbox.js';
 import { computeLocalWorkdir, computeRemoteWorkdir } from '../../sandbox/workdir.js';
